Combat Correspondent (#2761)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This adds the role combat reporter. This role will be a primarily RP
role. The role has an office south of shipside tcomms and west of
laundry.

For the time being they have the same restrictions as the CL:
They may deploy at will.
They report directly to the acting commander with regards to ML.
They are NOT to be fighting xenos or frontlining. They should be staying
in safe areas or moving with escorts (making the area safe).

# Explain why it's good for the game

This is something that has been asked for quite a bit. I'm on the fence
about it personally but was requested to get something running.

There is a good argument against "role bloat" which is generally
misunderstood as "too many roles bad" which is not the case. Instead,
role bloat involves creating extra roles that take away from the
gameplay loop or takes away from other roles (specifically in this case
the corporate liaison who is rather similar if they want to do a
gimmick).

Anyways, we'll see how the testmerge goes, feedback, as well as other
architect thoughts.

Oh and the feedback word today is GIGGLES.


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>
Tested a bit, we'll run into bugs I'm sure.
</details>


# Changelog

:cl: Morrow
add: Added the role Combat Correspondent.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: harryob <me@harryob.live>