Introduces a new way to rename items, and cassette pouches now take regulation tapes. (#2956)
# About the pull request
This implements a way of renaming items using the rename component. In
this case, it checks for a trait and if it has it the rename component
will name an item the initial item name with the new name in
parentheses. For example, if you rename a regulation tape "Interview 1",
it's gonna show up as "regulation tape (Interview 1)".

Tapes now also store the name for each side of the component, so if you
take the previously mentioned label of Interview 1 and then flip it, the
other side will be unlabeled until you alter it.

Also, regulation tapes now fit in casette pouches for use by the
reporter.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

This is a better way to handle tapes for tape recorders, makes things
more immersive for Combat Correspondents and is more convenient.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
add: Cassette pouches now fit regulation tapes.
add: Flipping a previously labeled tape will store the label for that
side and vice versa.
add: Tapes now have a unique renaming system.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->