Adds the Maintenance Jack (#2746)
# About the pull request

This PR gives the Maintenance Jack to ship synths in the experimental
vendor and Working Joes for free.
Gives the Maintenance Jack bonus effects
(Had to give WJs Engi 3 for one of the Maintenance Jack's abilities,
this shouldn't change much)

# Explain why it's good for the game

New tool for synths and WJs

# Changelog

:cl:
add: Added a Maintenance Jack for Synthetics (Experimental Vendor) and
Working Joes (Standard Kit). It has two modes, being able to switch
between a crowbar and a wrench. You can use it to unbolt doors, and, if
strong enough, pry them open.
/:cl: