Fix for jamming smartguns. (#3001)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Sometimes fixing a bug takes a few hours of searching through a mulitude
of files, seeding debug messages, tracing call stacks and trying to find
the root cause of the issue. Sometimes it is as simple as looking at the
one relevant file and going straight to the obviously responsible part
of it.
And every so often I spend a few hours on the former only to eventually
realise that all that was needed to be done was going straight for the
latter without assuming that _obviously_ it is not going to be that
simple because _obviously_ things that already work do already work and
nobody would break them. And then in hindsight I feel so stupid. But at
least the bug is gone.

# Explain why it's good for the game

Fixes #1870 and #2147.

# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
fix: Jamming smartguns now properly clean themselves (as they were meant
to be doing all along).
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->