Let (whitelisted) ghosts join in as Working Joes (#2963)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

https://forum.cm-ss13.com/t/make-it-so-ghosted-players-can-spawn-in-as-working-joes/588
Adds a ghost verb so ghosts can now join the game as working joes.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Like it was said in the forum post, Joes are pretty much harmless, their
impact on the round is pretty much minimal, so adding this wouldn't
imply any balance change whatsoever.

The intent of adding something like this to the server is mainly trying
to get more attention to the role. Which is barely played at the moment,
however, if dead players could now respawn as them, I believe it might
get more activity.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
Unsure of what kind of debugging I could provide in here, request any in
the comments and I'll do it.

There's a high chance that something is not treated as it should. This
is probably the first PR where I had to look and understand what I was
writing, despite the fact that a lot of the code is taken from the pred
initialization. So yeah, if you would have placed something somewhere
else, or differently, go ahead and tell me.

# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:H20Begod
add: Added a new verb for ghosts to join in as working joes (ONLY
WHITELISTED GHOSTS).
code: Edited max spawns of the working joes to 3.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->