New Steel Crest Ability: Soak (#2752)
# About the pull request

This adds a new ability for the defender strain, Steel Crest called
Soak. It is an ability that when activated, tracks damaged taken and
once it reaches a threshold slightly heal the Steel Crest defender and
reset the cooldown of its Tail Slam ability. The damage threshold needed
to reach this state is 140 damage within 6 seconds, and will not do
anything if it doesn't reach this cooldown.

# Explain why it's good for the game

Steel crest as a strain is extremely underwhelming, it's almost entirely
dedicated to simply soaking damage. Its combat capabilities is withheld
to headbutting from a corner and then going back in hiding. Its tail
slam is underutilized as a result, and this PR will switch up gameplay
and allow for steel crest to be a much more fun mix between its tanking
ability, and its damage dealing aggressiveness.


# Testing Photographs and Procedure

<details>
Tested against being hurt by a sentry while its active, it works. Will
refresh your tail slam and slightly heal.
</details>


# Changelog
:cl: Usnpeepoo
add: Added a new ability to the Steel Crest strain, called Soak. It
tracks your accumulated damage and once it reaches a threshold, reset
your tail slam cooldown and slightly heal you.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>