Goonify chemical retrieval (round two) (#3011)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR aims to do the following:
Replace the current PMC chem team with a corporate security chem team
led by a researcher.
The researcher can call a PMC team for assistance so we have escalation
if required. (admin approved)

This should allow for escalation of force for what should be a fairly
cooperative situation. Weyland Yutani and the USCM are close friends in
the Neroid sector. There is no reason this should blow up as much as it
does but there should be some tension.

As always, this is a work in progress. Nothing is 100% final.

# Explain why it's good for the game

Read above.


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>
I briefly tested some features of this.

</details>


# Changelog

:cl: Morrow
add: Chemical retrieval is now primarily corporate goons with a backup
team of PMCs if required.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: harryob <me@harryob.live>