Adding option doctor job to become a surgeon just for RP. (#3010)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
i am adding an option to become Surgeon as an only RP purpose.
remove most of the gear those jobs (nurse, doctor, research)spawn with
and put it in their respective vendor.

create a parent for medical preset to avoid to having three time the
same thing.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The idea is to show to other people that you are a doctor specialized in
surgery.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

change in character preference here you can choose between doctor or
surgeon.

![image](https://user-images.githubusercontent.com/117036822/229756969-66b22e66-8441-4738-b4d7-ec985c58c104.png)
Surgeon.

![image](https://user-images.githubusercontent.com/117036822/230122204-bb0216ea-4e3f-4539-8d97-e06d77f6855d.png)
Doctor.

![image](https://user-images.githubusercontent.com/117036822/230122331-8499f76d-b809-4c76-8765-72cbc411327e.png)

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
add: Added the option for the doctor job to become a surgeon so you can
choose to RP one.
add: Added the option in doctor vendor for medical belts, medical set,
surgical caps, doctor's snow vest.
del: remove most of the gear those jobs (nurse, doctor, research)spawn
with and put it in their respective vendor.
code: created a parent for medical preset to avoid to having three time
the same thing(back-end).

/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>
Co-authored-by: harryob <me@harryob.live>
Co-authored-by: Morrow <darthbane97@gmail.com>