LV522 Changes: Edits to LZ1 adding short fog along with removal of redundent area and addition of W-Y Vault (#3041)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR makes changes to the LZ1 area of LV522 along with the south
reactor flank being removed and the addition of a W-Y Vault as an
optional RP objective for corporate-aligned COs or greedy marines

This PR also gives the FORECON CO a Windbreaker jacket to make him more
recognizable to his men

# Explain why it's good for the game

LZ1: With my last LV522 PR I placed blast doors around LZ1 with the
intent being to deny survs and xenos from getting inside, this did not
work and I accidentally created a very strong hold area for FORECON this
PR aims to fix my blunder with the newly added short fog

South Filt: The area south of reactor filt was removed because I felt it
was redundant and made the map too large while adding nothing making it
harder for aliens to build in a centralized spot without fear of a
flanking route that completely bypasses it, a future planned reactor
rework aims to address multiple problems with LV522 but this is a
stopgap until then

At the request of nanu a new route has been added from north colony
command to south reactor filt

The Vault: I've added a W-Y vault north of Corpo dome east of fitness,
this area will likely not see use outside of the possibility of greedy
marines raiding it for a more corpo aligned CO wanting to secure the
money and gold inside

Tornado: I forgot to add this at first but the UD6 tornado has been
changed to require a button pressed from the Typhoon to open it up to
get the loot, this was added to require FORECON to move around more to
get good loot items instead of holding one spot

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:SpartanBobby
maptweak: added short fog around LZ1 on LV522 marines will not see this
as it will be gone by the time they land
maptweak: heavily edited a flanking route south of reactor filt on LV522
maptweak: Added a W-Y Corpo vault filled with gold and money to LV522
marines are encouraged to raid it Weyland aligned COs are encouraged to
secure it for their corporate overlords
maptweak: Removed some ways into the western reactor and some another
route lightly longer to get to
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->