Fixes the contents showing under crates/barrels when you disassemble them (#3043)
I hate this bug so much I don't know why it wasn't fixed until now it
makes maintenance/planetside looting so much more of a hastle than it
needs to be (you need to right click on everything you disassemble or
alt click the turf to see the stack? Who thought this was a good idea?)

The code takes the material of the crate and reorganizes the stack using
a new proc to ensure it is on the bottom of the list meaning that the
contents will show up on the top.


![dreamseeker_sm4yWDBgEP](https://user-images.githubusercontent.com/24533979/230979929-df577f6a-d611-431c-aff9-224361923140.png)


:cl: Hopek
fix: Disassembled crates/barrels will now correctly show their contents
above the deconstruction material meaning that you don't have to right
click or alt click the turf to see what you looted.
fix: Fixed duplicate crate/barrel disassembly material spawning because
it was being created in two places. If people want 2 material per
crate/barrel they'll have to intentionally code it in this time vs it
being a bug.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>