Removes Quickclot (#3040)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR removes quickclot. It is refactored into a property that can be
found in research.

# Explain why it's good for the game

Blood loss can and should be an important function of the game. While
debilitating when it gets bad, it *rarely* gets to that level. Having a
chemical that is standard in almost all kits that turns off what is
supposed to be a serious consideration medically is not interesting.

More blood changes coming.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Removed Quickclot
add: Added hemostatic chemical property.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>