Renames most instances of "Captain" to "Commanding Officer" in the COs clothing (#3084)
# About the pull request

Renames most cases of Captain in CO clothing and renames it to
Commanding Officer.
Fixes a typo in the XOs uniform description.

# Explain why it's good for the game

The CO role has not had the rank of Captain for a couple of years at
this point, updating the names to have the correct rank in them helps
avoid confusion as to who the items belong to.

Typos are bad and ugly and gross.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Foxtrot1322
spellcheck: Renamed CO items with "Captain" in their names &
descriptions to "Commanding Officer"
spellcheck: Fixed a typo in the XOs uniform description
/:cl: