Fixes the contents showing under crates/barrels when you disassemble them V2 (#3075)
Fixed edition of #3043

### Original PR description:
I hate this bug so much I don't know why it wasn't fixed until now it
makes maintenance/planetside looting so much more of a hastle than it
needs to be (you need to right click on everything you disassemble or
alt click the turf to see the stack? Who thought this was a good idea?)

The code takes the material of the crate and reorganizes the stack using
a new proc to ensure it is on the bottom of the list meaning that the
contents will show up on the top.


![image](https://user-images.githubusercontent.com/24533979/232187069-a0c074c5-6bd1-4483-8273-f952756620e4.png)

:cl: Hopek
fix: Disassembled crates/barrels will now correctly show their contents
above the deconstruction material meaning that you don't have to right
click or alt click the turf to see what you looted. Doesn't make the
contents break reality this time.
fix: Fixed duplicate crate/barrel disassembly material spawning because
it was being created in two places. If people want 2 material per
crate/barrel they'll have to intentionally code it in this time vs it
being a bug.
/:cl:

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>