Gives FORECON Synths red shoes the ability to hold boot knifes (#3055)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR adds redshoe/knife a subtype of the red shoe that allows a knife
to be stored in it for use by the FORECON synth at this time.

# Explain why it's good for the game

After getting feedback from the synthetic senator about problems with
the FORECON synthetics equipment the lack of ability to store a knife in
their shoes was brought up and since the FORECON synth is the only
equipment set to have this problem I figured I'd sort it out

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:SpartanBobby
add: FORECON Synthetic now has a special type of red shoe with the
ability to hold a knife
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->