Resin whisperer rework (#3094)
Continuation of #3089

This rework is mainly focused on moving the Resin Whisperer to a more
focused role of pushing the front-lines with his team.
This needs a test merge to see if people like it.

Here is a summary of the major changes:
1. The Resin Whisperer can no longer use the Coerce Resin ability when
not standing on weeds.
2. The Resin Whisperer can only place resin nodes on existing weeds to
slowly "push" weeds outward versus just coating a large unwedded area on
their own. **(I promise this is leading somewhere and isn't just nerfs
bear with me please)**
3. The Resin Whisperer whisperer can no longer plant weeds normally; the
ability has been removed **(please just a little while longer)**
4. The plant weeds ability for the Resin Whisperer has been moved to a
construction in the Resin structure tab. This means that it must now be
built using Coerce Resin. This means that the Resin whisperer can now
remotely build resin nodes. This requires them to channel the
construction as per normal and costs substantially more than it did
previously.
5. There is a warning displayed to the marines when the Resin Whisperer
is going through the channeling process to build a Resin node on the
exact tile the resin node will appear.
6. The resin whisperer can "refresh" a resin node meaning that they can
build another resin node on top of an existing one to restart the
spreading process if the node stopped growing and there is spaces to
fill (this is how you get past doors). Refreshing a node resets it back
to low health to grow over time.

**Keep in mind that these resin notes start off weak as per the usual
mechanic and gain HP over time to the usual amount we are used to so a
marine can literally walk up to them to stab them in a single hit to
remove them.** We normally don't see this in gameplay because there is
usually an alien right on the nodes being placed so a marine doesn't get
a chance to attack the node quickly after it being placed.

What will possibly be changed depending on how the test merge plays out:
- Currently the resin node cost for the Whisperer is 2x the cost of a
thick membrane. I can see this changing to 3x if they're able to place
too many of them.
- Longer casting time for the channeling of the resin nodes
- Axing the whole rework entirely because the community will be outraged
and send bitcoin assassins to my home



![dreamseeker_ugCZ52AeUN](https://user-images.githubusercontent.com/24533979/232867239-7ade0c29-9d0a-471e-b652-460e6eb5bd33.gif)

"But how do I spread past a door when resin nodes stop spreading over
time?"
Just refresh a node like in this example below. Keep in mind that the
node starts will low HP again meaning it'll be 1 hit away from death


![dreamseeker_jOi3IveIu7](https://user-images.githubusercontent.com/24533979/232867667-6de7af74-a254-4855-8292-e3a5d32d2368.gif)

Another GIF because why not:

![dreamseeker_qLSICX4u7Y](https://user-images.githubusercontent.com/24533979/232870258-920f3ebd-00a9-4af2-b3f0-cca43e7cd189.gif)



# Changelog

:cl: Hopek
add: Resin Whisperer has been reworked. Major changes include being able
to place resin nodes remotely on weeds, not being able to use coerce
resin when standing outside of weeds and losing the ability to regularly
plant weeds outside of current existing weeds. This changes the Resin
whisperer to a new "slow crawl" frontline advance gameplay where he
assists his team in pushing the front lines forward as he is able to
both advance weeds and build structures on them.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Segrain <Segrain@users.noreply.github.com>