Fix predaliens being unable to attack enemy xenomorphs
Removed "iscarbonsizehuman" check which was preventing the predalien from doing damage almost every single type of enemy xenomorph as all of them have a mob size higher than that of humans.
Simplified the code that checks if the attacked character is a yautja, which was also previously preventing the predalien from attacking enemy xenomorphs when isolated from the other check.