mapping renovate the medical lobby (#3054)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

1-Improve the medical storage so that the medical staff can distribute
supplies while inside of it.
2-remove the walls around the reception desk and align the door with
those east.
this helped a lot to gain space to make the medical storage waiting
line.
3- added a bunch of empty medickit.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

1-allow medical personnel to more easily distribute supply to marines.
2-make it easier for the one at reception to help patients.
3-avoid clustering of the floor in medbay.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>


![image](https://user-images.githubusercontent.com/117036822/231552538-445ca44b-39b4-4d4f-8ef0-1be4d3b50bbf.png)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
maptweak: shuffle items in medical storage to make it possible for
personnel inside to distribute medical supplies.
maptweak: reduce the size of the reception office and remove is wall and
window to make it so that personnel inside can act quickly when patient
arrived.
mapadd: added multiple empty medikit to avoid marines clustering the
floor when making their loadout.
maptweak: move all the empty pill bottle in a closet to a spot in the
chemistry room.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>