Marine Major announcement delay (#3118)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR makes it so when a marine major occurs (no xenos left) it will
hold for X minutes before it does the "credits" and song and dance.

The way I did this is probably cursed.

Time open for discussion. Currently at three minutes.

# Explain why it's good for the game

This should allow for a bit more RP time at the end of the round and
potentially some time for medals, congratulations, etc.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Added delay to end of round announcement when it is a Marine Major.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: naut <55491249+nauticall@users.noreply.github.com>