Adds a Spec to the Foxtrot Cryo Squad. (#3151)
# About the pull request
Adds a Squad Weapons Specialist to the Foxtrot ERT bought through
DEFCON.

# Explain why it's good for the game
Foxtrot as it currently stands is in a poor spot, hence the meme
"Foxtrot gets called, it's 3 guys, 2 of which cryo immediately",
hopefully the inclusion of a Foxtrot Spec will increase the number of
people signing up for it, as it is a pretty big incentive.

# Testing Photographs and Procedure
Compiles and runs, ERT can trigger fine, still.

# Changelog

:cl:LynxSolstice
balance: Adds a Squad Weapons Specialist to the Cryorines.
/:cl:

---------

Co-authored-by: Morrow <darthbane97@gmail.com>