Radio backpack tweaks and fixes (#3164)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR allows for making phone calls or answering phone calls while
holding the backpack in hands.

This PR stops the shake verb from shaking non-items out of storages.

This PR fixes a content offset for non-items in storages. This fixes the
off by one error when trying to grab things out of a radio backpack.
Fixes #2233

More radio backpack stuff to come as motivation waxes and wanes.


# Explain why it's good for the game

With the upcoming removal of RTO uniform requirements for the backpacks
they should be made easier to use outside of needing to be on your back.

Bugs bad.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Radio backpacks can now be used in hand
fix: Fixed the off by one error in radio backpack item removal
fix: Shaking a bag no longer drops non-items
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->