Ship vs Ship - base verb (#3015)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

GENERAL QUARTERS. ALL HANDS, MAN YOUR BATTLE STATIONS.

Unprofessional, but get's it through!

This is the first PR about this to be added from the ideasguys post that
I made. I have a feeling I absolutely cannot do the other stuff though,
but I hope this is already a really interesting addition!

This adds a whole new section of the admin tab for the staff to play
with.
It will contain the main verb to shoot sticks of boom at the Almayer, a
small button to call General Quarters on the Almayer at the request of
the commander, for dramatic effect, and the Shakeship verb will also be
moved here when the rest of the things are already in.

# Explain why it's good for the game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

It's basically just for events. But will make the whole experience MUCH
smoother.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
admin: Added a weapons panel in a ship section of the admin tab, where
you can shoot weapons at the Almayer.
admin: Added a General Quarters button to the ship section of the admin
tab, where you can call General Quarters if the ship's in combat.
admin: changed where shipshake is, to be under the ship section.
soundadd: Added general sounds related to the receiving end of a ship vs
ship encounter, with two different General Quarters call, one containing
the traditionally used Bosun "All Hands" call(Which is the one used in
the General Quarters verb). For tradition.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>